Added definition of the core client object
I created a core client object that actually dials up to the Fewer Service server app, then performs the GetAggregatesStream RPC and logs in its client logs the activity going on during the operation.  This object would be used in different implementations of the Fewer Service Client Application (e.g., a CLI, a client object used at different times by a microservice calling on the Fewer Service, etc.).  The PerformGetAggregatesOp method is separated from the methods for ConnectToServer and Close, so that way if some implementation desired to call on the core RPC operation multiple times, it can do so with ease.