Simplified client logger object definitions
During the writing of the server logger object definitions, I realized that the development-level logging objects -- both client and server side -- were simply extensions of the production-level loggers because they use the info, warn, and error level loggers, which are also methods that satisfy the general ServerLogger interface.  Therefore, I decided to change the definition of the development-level logger to simply embed the methods of the production-level struct.  There were no errors in making this change.