Added general gRPC server for Server Application
Now that the core Fewer Service operations were defined, I added the definition of the general gRPC server that will use a given net.Listener object to listen to all requests at a given address and port, create a server logging object based on whether the server is expected to be in a production or development/test environment, and register an instance of the core Fewer Service and a side gRPC reflection service to the server.  Once the server is setup and listening for requests, it is also listening to any potential OS termination or interruption signals, so that in the case something causes termination or interruption, the general server can shut down the service gracefully and close any resources the server logging object uses during activity logging.