redefined the client and server loggers for ordered logging
During some tests, I realized that the server terminal logs were logging events randomly, not in the order in which they actually happened, when the 'go' keyword was being used on the terminal logger lines.  Therefore, I removed the keyword from all those lines.  This should allow for ordered logging.