Added flag parse to main server setup code
Realized that the main server setup code did not have a call of flag.Parse() to actually load and parse values of 'go run' command flags, which was why the server_devtest.log file was not being created or appended to.  This situation has been fixed, and now users can easily configure server to be dev by using --prod=false flag.